debugged service methods and changed models from tests

+++ internal/features/users/user_service_interface.go
@@ -5,5 +5,5 @@ type IUserService interface {
 	GetUser(id int64) (User, error)
 	GetUsers() ([]User, error)
 	UpdateUser(update User) error
-	DeleteUser(id int) error
+	DeleteUser(id int64) error
 }
